GLOW-14 Zero-fill missing data

+++ src/main/java/ru/rbs/stats/data/merchants/SQLReportBuilder.java
@@ -48,7 +48,7 @@ public class SQLReportBuilder implements StatsReportBuilder {
                 for (String aggregateName : config.getCubeDescription().getAgregates()) {
                     switch (config.getCubeDescription().getType(aggregateName)) {
                         case INTEGER:
-                            entry.getAggregates().add(rs.getInt(aggregateName));
+                            entry.getAggregates().add(rs.getLong(aggregateName));
                             break;
                         case FLOAT:
                             entry.getAggregates().add(rs.getFloat(aggregateName));
